fix(catalog): Withdrawn "top" controls displaying
This fixes the "state" of withdrawn "top-level"
controls. Instead of being a collapsible item
that shows a blank box when selected, the
control no longer opens and closes, as
indicated by no arrow appearing to the far
right of the item. The item text is also styled
to appear a light-gray and has a strikethrough,
like the "sub-level" controls.