On branch mpemy/display-remove-control-elements  Your branch is up to date with 'fork/mpemy/display-remove-control-elements'.
 Changes to be committed:
        modified:   src/components/OSCALControlProse.js
Fixing the failing build due to linting issue.
The line of code that was causing this issue have been addressed.
There was method MarkupLines that was  added and then removed because it was not needed
I just forgot to delete it that is why lint was failing because I was importing a component that was not used and was inexistent.
I have simply deleted that reference.