fix(catalog): Withdrawn "top" controls display correctly (#720)
This fixes the "state" of withdrawn "top-level"
controls. Instead of being a collapsible item
that shows a blank box when selected, the
control no longer opens and closes, as
indicated by no presence of an arrow
appearing to the far right of the item. The
item text is also styled to appear a light-gray
and has a strikethrough, like the "sub-level"
controls.

Resolves #635

---------

Co-authored-by: Kyle Laker <klaker@easydynamics.com>