fix(control): controls display id instead of label
Controls would display the `id` field a lot except for the like one
place where we did properly lookup the label. This adds a new generic
way to query for props by name (ignoring value). This retains `id` as a
fallback as a bridge way to move to identifying whether that is better
or if we should instead just not display anything label-like at all. The
latter requires far larger refactoring.