fix(control): controls display id instead of label (#746)
Currently, controls are often displayed with their `id` in uppercase
instead of using the `label` property. This means that the display is
not necessarily always what a document author may intend (even if it
perhaps is close).

The label is displayed matching the surrounding content. If the ID is
used, it falls back to displaying using the ID (in our standard grey
color).

This also fixes an issue where, for prose, the label was squished up
against the text without a space.


### Screenshots - [Test
Catalog](https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/test-content/catalogs/basic-test-catalog.json)

#### Before


![image](https://user-images.githubusercontent.com/850893/228004046-99870901-fa7c-42a8-8c13-0cdaca4c0e9d.png)

#### After


![image](https://user-images.githubusercontent.com/850893/228005449-6794e17e-a420-493d-89d4-3d74471e1fb5.png)

### Screenshots - [SP
800-53](https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/main/catalogs/NIST_SP-800-53_rev5_catalog.json)

#### Before


![image](https://user-images.githubusercontent.com/850893/228030191-9153c84d-87c5-4c66-bfd7-984d6ae4bc04.png)


#### After


![image](https://user-images.githubusercontent.com/850893/228030062-8c9d4300-837e-41e6-90bf-10c4b9a4c620.png)