refactor: switch to a monorepo with lerna (#722)
This refactors the repository into a more typical monorepo setup. This
will eventually allow to add additional packages (perhaps an API client
or a package that exports OSCAL types for TypeScript). Along with this
refactoring, we've replaced `create-react-library` with `rollup` for the
React library and we use `craco` instead of raw `react-scripts` in the
Viewer app (to workaround some wonkiness with forbidden imports outside
of `src/`).

Along with this we get TypeScript!! Only one or two files have been
rewritten as part of this migration. That serves two purposes, first
that we can use TypeScript, and second that we can adopt it
incrementally.

We also have a `scripts/` directory which will allow us to store scripts
that we want that are too complex to just toss into `package.json`. I
would like to keep those scripts limited to Bash (all our users should
have access to a basic Bash environment) or to TypeScript (we install
`ts-node` as a dev dependency in the root repo).

`lerna` is used as the tool to manage our monorepo. A lot of features
are now built into the `npm` workspaces features; however, `lerna` gives
us some nice ability to manage references between repos and to run
commands against all of them (and look at how nice that is for
`scripts/watch-viewer`!)

The tests are all still `jest` and our linting rules remain largely the
same. Tests should be run in the individual packages for now;
functionally for the majority of our work, this means that you need to
be in `packages/oscal-react-library` in order to run its tests for
example. For commands that do work across all packages without any
dependencies (for now, just `lint`) those can be run across all packages
from the root.

If you do want to run across all packages, `lerna exec --stream` is a
good friend; additionally, its `--scope` and `--include-dependencies`
and `--include-dependents` arguments may be good friends.

You _do not_ need to install `lerna` globally (although you may find it
useful to do so). You can use `npx lerna` (as the example in
`scripts/watch-viewer` does).

Closes: #732 
Closes: #403
Closes: #596 
Closes: #493