chore(lint): allow lines up to 100 characters (#766)
This is a pretty trivial linter change but with the addition of
TypeScript, lines for function definitions and elsewhere are going to
get a bit longer. This gives us some more room on each line before they
get split unnecessarily early.

The true changes here are:

 - updating `.prettierrc` for the new line length
- updating `packages/oscal-viewer/package.json` to properly reference
the config file

The remaining changes are just the result of `npm run lint:fix`.

After this is merged and we have the commit hash, I will update
`.git-blame-ignore-revs`[^1] with that to clean up `git-blame(1)` output
in the future.

[^1]: While
https://github.blog/changelog/2022-03-24-ignore-commits-in-the-blame-view-beta/
isn't the formal documentation for this feature, it covers the intent of
the file. The same file is supported by the `git` tool locally.