feat(back-matter): edit backmatter resources
This adds basic editing functionality for backmatter resource
titles and descriptions.

Our original editing functionality made some assumptions that made
implementing this tricky. Namely, we assumed that the text field we are
updating is the direct child of the field we are sending a PATCH request
for.

For example, `metadata.title` updating `metadata` vs
`backmatter.resources[i].title` updating all of `backmatter`.

As you can see in the code, the way we handle editing is a bit hacky and
naive. We most likely want to revisit this and redo a lot of that code.
This would not be an insignificant change, hence it being pushed for
now. These changes make editing at least work for now so that our other
editing tasks are no longer blocked.