feat(documentation): Document how to add editing
This attempts to document how one would go about adding
editing to a text field.

This **does not** mean this is the best way to actually do this.
To clear up the blocker of editing, this at least makes it possible to
edit. We will want to revisit editing at some point in the future.