feat(loader): allow processing local files
This allows users to "upload" a file to the application without having
to publish it on the web. The user just needs to choose "upload" and
point to the file on disk that they would like to use. This creates an
object URL in the browser (so the file is not actually uploaded
anywhere).