feat(loader): allow processing local files (#808)
This allows users to "upload" a file to the application without having
to publish it on the web. The user just needs to choose "upload" and
point to the file on disk that they would like to use. This creates an
object URL in the browser (so the file is not actually uploaded
anywhere).

Using a web URL

![image](https://user-images.githubusercontent.com/850893/234646794-9aeee5c5-1765-4964-89a0-19b4b8140d88.png)

Using a local file

![image](https://user-images.githubusercontent.com/850893/234648859-a06e4783-ecf1-45ac-9155-3fa344e08c32.png)

The URL is automatically generated by the browser. This is the best
compromise to allow "Reload" to work as expected.