chore: add types to the loader (#830)
This takes a first swing at adding type information *making as few
changes to the code as possible* barring very explicit bugs or things
required to get types to work.

There is a non-exported and deprecated `Mutable` type added in the REST
Utils. It should not find any additional use and all existing uses
should eventually be replaced. This is an example of a bad thing done to
avoid breaking code.

Additionally, I don't understand how `populatePartialRestData` ever
worked. There is all kinds of stuff with checking `.length` and then
trying to use that list (or single-character string?) to index a list.
That is the biggest functional change to the code here.

I would appreciate knowing this has been tested against various editing
paths in the editor. If you reproduce a bug, be explicit about what you
edited and include the built path, request, and any response data (and
logs from the server) and post it as a comment on the PR.