chore: use FedRAMP SSP template as the default SSP
There are a few benefits to making this change. Primarily, this is a
much more complete SSP than our existing example, despite the fact that
it is a template. This gives better visibility into how many fields in
an SSP are/may be used. Additionally, it proves that our viewer works
directly with content that we did not curate specifically in our own
repository.

This make break tests that rely on the content of the SSP to be a
specific value (but if they want that, they should provide a specific
SSP).