fix: make properties button smaller
This changes the "Open Properties" button to be only an icon. This then,
in turn, makes it so we can put it in the Control Title areas and in the
action areas for Back Matter resources. This should give us more
flexibility about where and how to display this without "Open
Properties" as text all over the page.

A workaround is added to make it so that even a `disabled` properties
button appears with a tooltip; this should make it easier to understand
the purpose of the button. As explained in the inline comment, we can
probably remove that when we add full editing support.