fix: make properties button smaller (#840)
This changes the "Open Properties" button to be only an icon. This then,
in turn, makes it so we can put it in the Control Title areas and in the
action areas for Back Matter resources. This should give us more
flexibility about where and how to display this without "Open
Properties" as text all over the page.

A workaround is added to make it so that even a `disabled` properties
button appears with a tooltip; this should make it easier to understand
the purpose of the button. As explained in the inline comment, we can
probably remove that when we add full editing support.

Finally, the tests are reworked to centralize testing in a single file
and
to actually test the contents and different properties there. Tests are
left
in `OSCALMetadata.test.tsx` to have a test for a consumer of the API.


![image](https://user-images.githubusercontent.com/850893/236334681-bf7883ef-f6a1-4771-ac84-e948daf58199.png)