feat(control): uniform display for control parts (#849)
This makes it so that Discussion, Assessment Methods, and Objectives are
laid out similarly. Overviews and Statements always appear at the top
(in that order); all other parts are grouped within the drop downs. This
gates them behind a click (Discussion already was) but it makes the
controls a bit easier to read without a wall of Assessment Objective
text.

Component Definition: 

![image](https://github.com/EasyDynamics/oscal-react-library/assets/850893/b86800ab-14da-4cb5-af4a-29256d0ca2b2)

Catalog: 

![image](https://github.com/EasyDynamics/oscal-react-library/assets/850893/bec50bd0-881c-4a6d-b3b6-b5550832de27)