fix: large control display "squishes" tabs for Component controls (#855)
Historically, our display of controls for components has been a bit
fragile, where if the control side got too large, it would squish into
the component tabs and cause overflow, truncation, or wrapping. This is
suboptimal and instead we should use a behavior more like what we do in
OSCALControlGroups where we set a specific side on Grid items. The same
numbers are used in both places.

Because of the complexity around handling fragments, I did not rework
this into a shared component. No other styles for the component have
changes to align it with Controls (though we should do that).