feat: create hoverable popover (#853)
This creates a [popover](https://mui.com/material-ui/react-popover/)
component that is displayed on hover over. The main component is
intentionally fairly generic. You can imagine many different components
you would want to hover over and see text for.

There are also two wrapper components created for the common use cases
today of displaying `MarkupLine` and `MarkupMultiLine` in a popover.
This has been implemented in `OSCALSystemImplementationComponents` as an
example.

closes #592 

### Testing

- The default SSP and our old default SSP contain components with a
description to display on hover.

---


https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/main/system-security-plans/ssp-example.json


![image](https://github.com/EasyDynamics/oscal-react-library/assets/39490074/4cc61d83-81e8-49cf-ba62-6f545b44e55a)

---------

Co-authored-by: Kyle Laker <klaker@easydynamics.com>