feat: human-readable display for back matter media types (#861)
This updates how we display resources in Back Matter to try to use a
more human-readable name. When `media-type` is provided, we prefer that;
otherwise we fallback to guessing based on the file extension using the
`mime-db` package's database (which includes registrations with IANA as
well as the Nginx and Apache databases).

The media types are largely those defined in #111 though with some
additional ones found during the implementation that may be useful. The
lookup algorithm is not significantly modified besides the feedback
given in the issue comments.

The UI is also updated slightly to give more visual indication that it's
a button and can be clicked to open the link.

Closes: #111

### Screenshots

General Display:

![image](https://github.com/EasyDynamics/oscal-react-library/assets/850893/23418822-391c-4106-9280-db6af28d2a2a)

OSCAL Media Types:

![image](https://github.com/EasyDynamics/oscal-react-library/assets/850893/750a9502-35f2-43bd-8709-32c4b77dde48)