feat: display params in catalogs, groups, controls (#858)
Creates a dialog box with accordions to display parameters. In theory,
there can be quite a bit of information put here so accordions were
chosen over using a table like we do for properties.

closes #822 

### Testing

Based on example catalogs today, often most fields are undefined and
only a few things are actually displayed. This example shows a "worst"
case parameter definition.


https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/tuckerzp/params/catalogs/basic-test-catalog.json

---


![image](https://github.com/EasyDynamics/oscal-react-library/assets/39490074/978f789f-6bdc-4e77-ac45-6361f620005b)

---


![image](https://github.com/EasyDynamics/oscal-react-library/assets/39490074/ba847f15-b195-4a6c-a2bc-89225cb82611)


![image](https://github.com/EasyDynamics/oscal-react-library/assets/39490074/8603a4d2-ba66-4b1f-93f6-c196651de790)

---------

Co-authored-by: Kyle Laker <klaker@easydynamics.com>