fix: update type workflow fails on v1.1.0 release (#916)
The v1.1.0 OSCAL release's ZIP file doesn't have a "wrapping" folder
named for the release version like previous versions did. Basically,
this means that when you explode the ZIP, you get the `json` and `xml`
directories directly (instead of having them inside an `oscal-1.1.0`
folder as in the past). This change checks to see if a `json` folder
exists in the current directory; if so, it looks in there for the
complete schema. Otherwise, it looks for an `oscal-VERSION` folder. If
neither exists, it's a new case we haven't seen before, so there will be
an error anyway.